Lower default value for broker_commit_every to 10
This should prove the theory that the config arg is not being consumed and only taking the default.
Despite passing the config arg with a value of 2 we only seeing the offets committed every 20k in rawstream (2 pods x 10k).
Our faust agents have been highly unstable and it's very uncommon for them to successfully process 10k messages before getting restarted.
If we see the offsets regularly drop after this change we will confirm the config arg is not being properly consumed